OR-0008 - move MockVideo to new model dir, ui for downloading exception
