+++ app/src/main/java/com/silverorange/videoplayer/repository/IVideosRepository.kt
@@ -1,6 +1,6 @@
 package com.silverorange.videoplayer.repository
 
-import com.silverorange.videoplayer.MockVideo
+import com.silverorange.videoplayer.model.MockVideo
 
 interface IVideosRepository {
     suspend fun fetchVideos(): List<MockVideo>?
